refactor: Rename RacingGame class
