+++ src/main/java/racing/controller/RacingGameController.java
@@ -2,10 +2,9 @@ package racing.controller;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import racing.RandomNumberGenerator;
 import racing.domain.Car;
+import racing.domain.RacingGame;
 import racing.dto.GameResultDto;
-import racing.service.RacingGameService;
 import racing.view.InputView;
 import racing.view.OutputView;
 
@@ -20,12 +19,12 @@ public class RacingGameController {
                 .collect(Collectors.toList());
         int tryCount = inputView.inputTryCount();
 
-        RacingGameService gameService = new RacingGameService(new RandomNumberGenerator(), tryCount, cars);
-        while (!gameService.isEnd()) {
-            GameResultDto resultDto = gameService.play();
+        RacingGame game = new RacingGame(tryCount, cars);
+        while (!game.isEnd()) {
+            GameResultDto resultDto = game.play();
             outputView.printRacing(resultDto);
         }
-        List<String> winnerNames = gameService.getWinners();
+        List<String> winnerNames = game.getWinners();
         outputView.printWinners(winnerNames);
     }
 }
